This commit was manufactured by cvs2svn to create branch 'mini_gviz'.
git-svn-id: http://svn.code.sf.net/p/genoviz/code/branches/mini_gviz@2236 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genoviz_sdk/src/com/affymetrix/genoviz/widget/neoseq/WrapStripes.java
@@ -95,8 +95,7 @@ public class WrapStripes extends WrapGlyph {
 
           int yoffset =
             stripe_start % (stripe_colors.length * residues_per_stripe);
-          int stripe_num =
-            (int)(stripe_start/(residues_per_line * residues_per_stripe));
+          int stripe_num = stripe_start/(residues_per_line * residues_per_stripe);
 
           backCoordPoint1.x = 0;
           backCoordPoint1.y = stripe_start;
